add pr data

+++ src/HubSync/Models/Repository.cs
@@ -5,7 +5,7 @@ namespace HubSync.Models
     public class Repository
     {
         public int Id { get; set; }
-        public int GitHubId { get; set; }
+        public long GitHubId { get; set; }
         public string Owner { get; set; }
         public string Name { get; set; }
